requirements.txt
# Telegram Bot
python-telegram-bot==20.7

# Database
SQLAlchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Task Scheduling
APScheduler==3.10.4

# Logging
structlog==23.2.0

# Configuration
python-dotenv==1.0.0

# Utilities
pytz==2023.3

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==24.3.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0
isort==5.13.2
